fix -Wnarrowing build warnings

+++ RFFC5072Lib/rffc5071.c
@@ -129,7 +129,7 @@ void rffc5071_setup(rffc5071_driver_t* const drv)
 static uint16_t rffc5071_spi_read(rffc5071_driver_t* const drv, uint8_t r) {
 	//(void)drv;
 
-	uint16_t data[] = { 0x80 | (r & 0x7f), 0xffff };
+	uint16_t data[] = { (uint16_t) (0x80 | (r & 0x7f)), 0xffff };
 	spi_bus_transfer(drv->bus, data, 2);
 	return data[1];
 }
@@ -137,7 +137,7 @@ static uint16_t rffc5071_spi_read(rffc5071_driver_t* const drv, uint8_t r) {
 static void rffc5071_spi_write(rffc5071_driver_t* const drv, uint8_t r, uint16_t v) {
 	//(void)drv;
 
-	uint16_t data[] = { 0x00 | (r & 0x7f), v };
+	uint16_t data[] = { (uint16_t) (0x00 | (r & 0x7f)), v };
 	spi_bus_transfer(drv->bus, data, 2);
 }
